Bug 2049856 - Remove old viaduct backend code
* Removed the old `backend` and `backend::ffi` modules
* Renamed `new_backend` to `backend`.

Changed the default timeout for `viaduct::Client` to 7 seconds for iOS
and 10 seconds otherwise.  This matches the behavior of we had for
`Request::send()`.  Should we make these longer?

Also:
 * Fixed build errors when crates depend on viaduct with the ohttp
   feature, but not rusqlite.
 * removed viaduct-reqwest which was no longer used. 
 * Removed un-used protobuf code.
