--- components/support/viaduct-reqwest/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "viaduct-reqwest"
-version = "0.2.0"
-authors = ["Sync Team <sync-team@mozilla.com>"]
-edition = "2021"
-license = "MPL-2.0"
-
-[lib]
-crate-type = ["lib"]
-
-[dependencies]
-error-support = { path = "../error" }
-viaduct = { path = "../../viaduct" }
-reqwest = { version = "0.11", features = ["blocking", "native-tls-vendored"] }
-once_cell = "1.5"
