+++ components/viaduct/Cargo.toml
@@ -18,12 +18,19 @@ serde_json = "1"
 once_cell = "1.5"
 parking_lot = { version = ">=0.11,<=0.12" }
 pollster = "0.3.0"
-prost = "0.13"
 ffi-support = "0.4"
 thiserror = "2"
 uniffi = { version = "0.31" }
 bhttp = { version = "0.7.2", optional = true }
 ohttp = { version = "0.7.2", features = ["client", "server", "app-svc", "external-sqlite"], default-features = false, optional = true }
+
+# Needed because we set the `ohttp/external-sqlite` flag which means that ohttp expects cargo to
+# arrange for it to be linked to `libsqlite3`.  This means we need to depend on rusqlite with the
+# `bundled` feature to bring it in.
+#
+# Without this, builds will often succeed because other crates will bring in `rusqlite`, however
+# some combinations will fail.  `cargo -p ads-client -p context_id` is one example (2026/06/25).
+rusqlite = { version = "0.37.0", features = [ "bundled" ] }
 
 [features]
 default = []
