+++ examples/viaduct-cli/Cargo.toml
@@ -11,8 +11,7 @@ serde_json = "1"
 url = "2"
 viaduct = { path = "../../components/viaduct", features = ["ohttp"] }
 viaduct-hyper = { path = "../../components/support/viaduct-hyper" }
-viaduct-reqwest = { path = "../../components/support/viaduct-reqwest" }
-clap = {version = "4.2", default-features = false, features = ["std", "derive"]}
+clap = {version = "4.2", default-features = false, features = ["std", "derive", "help"]}
 cli-support = { path = "../cli-support" }
 
 [features]
