+++ megazords/ios-rust/src/lib.rs
@@ -29,4 +29,3 @@ pub use tabs;
 pub use tracing_support;
 pub use viaduct;
 pub use viaduct_hyper;
-pub use viaduct_reqwest;
